update to new `withReduxSaga`
I've refactored and reworked `withReduxSaga`, `store/index.js` and the
various saga files to come up with a way to do `withReduxSaga` as a
possible npm package.

Everything needed for `withReduxSaga` to work is confined to the
`withReduxSaga` directory.

The `store/index.js` needed to change the export from a function to an
object that holds the `configure` function and the array of `rootSagas`.
The `rootSagas` are used internally by `withReduxSaga`.